+++ biopipen/reports/bam/CNAClinic.svelte
@@ -1,0 +1,9 @@
+<script>
+    import { Image } from "@@";
+</script>
+
+{% for gwpng in job.out.outdir | glob: "*/*.png" %}
+{%  set sample = gwpng | dirname | basename %}
+<h1>{{sample}}</h1>
+<Image src="{{gwpng}}" />
+{% endfor %}